Move shape parsing to avenger crate and update renderer to use it
