+++ avenger-vega/src/marks/path.rs
@@ -1,14 +1,15 @@
 use crate::error::AvengerVegaError;
 use crate::marks::mark::{VegaMarkContainer, VegaMarkItem};
-use crate::marks::symbol::parse_svg_path;
 use crate::marks::values::CssColorOrGradient;
 use avenger::marks::mark::SceneMark;
 use avenger::marks::path::{PathMark, PathTransform};
 use avenger::marks::value::{ColorOrGradient, EncodingValue, Gradient, StrokeCap, StrokeJoin};
+use avenger::marks::symbol::parse_svg_path;
 use lyon_extra::euclid::Vector2D;
 use lyon_path::geom::Angle;
 use serde::{Deserialize, Serialize};
 use std::collections::HashSet;
+use avenger::error::AvengerError;
 
 #[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize)]
 #[serde(rename_all = "camelCase")]
@@ -140,7 +141,7 @@ impl VegaMarkContainer<VegaPathItem> {
             let paths = path_str
                 .iter()
                 .map(|p| parse_svg_path(p))
-                .collect::<Result<Vec<_>, AvengerVegaError>>()?;
+                .collect::<Result<Vec<_>, AvengerError>>()?;
 
             mark.path = EncodingValue::Array { values: paths };
         }
